Replace RB_TREE with avl tre in lru.
It appears OS X does not have an easily available sys/tree.h. It seems
the easiest solution to this is to use our existing AVL tree instead, it
was a quite simple switch. I haven't checked how the performance change
but we know the AVL code is fast so I expect it to be close to zero at
worst.

Since lru was the only code using tree.h I've removed it.