Add tests for the cgo code.
To test this properly, I added an output tag to the xUnit testcase
output. This is not in the schema but is an extension. We already also
have another <log> tag that's also not in the schema.

Also, ninja output and go test output have to be verbose so grep can see
that the cgo tests are run.