Make plog fallback logging more nice.
Setup now detects if the fallback writer is a tty (typically stderr). If
so it uses a format similar to previously, but slightly improved. The
output is now JSON encoded for example.
If it's not a TTY, JSON objects are logged instead, using a format
similar to plogd. It's not identicaly however, since we need to log
directly and can't aggregate into objects. Thus a "key" field is added,
telling where the message would have been put by plogd.