Add a bunch of convenience functions.
Partly based on the logrus package, allow for the WithFields syntax it
has. LogDict can however be used to accomplish the same thing in a more
"plog" way. But now both works.

I guess this drops the thought I've had that the standard logging level
messages should only be strings. But "only strings or a dictionary with
a msg key that's a string" kind of works as well I suppose.

I used the previously unadded test context for plog here, so that's
added as well.