Add a configuration truct for fd pool services.
This is an alternative to bconf, a better one. As such I deprecated the
old AddConf name since it's confusingly similar.
I ended up adding this to cpool as well as I wanted it in the common
interface. That exposed a bug in the vtree_literal code that it couldn't
handle a keys_and_values with non-empty prefix. Most of that change is
indentation though.