Replace etcd client package with a lightweight internal one.
The coreos package lives in the main etcd module which has a lot of
dependencies. Sebase should have a low set of dependencies and therefore
I decided to remove the use of that package.

For now, tests are deferred to the packages using the new etcdlight
package. Watch-service was previously not tested at all, so I added a
simple test for it in the sapp tests.