Split the plog Logger interface into Logger and TypeLogger.
Technically this is backwards incompatible, but only as long as you're
trying to implement the Logger interface yourself, and I doubt anyone
does that. I added a disclaimer comment regardless that the interface
might change further in minor versions.

The TypeLogger is similar to Logger but it has the type key pre-set so
the methods don't accept that value.

Another part of this commit is the addition of the With method to
Logger, TypeLogger and Plog. It allows to store more values into the
interface types.