Add bufio on all sides of plog.
On the server side, this is just when writing to a file for now.
The network writer has its own buffering, maybe they should be merged
later.

After these changes profiling shows less then half time spend in plog
functions compared to before during a load with heavy logging.

After this change the underlying writer/reader for protobuf writers/readers
are no longer available, but I think that's sensible. Bufio.Writer.Write
is not thread safe so have to switch to a mutex for protecting it.