Add periodic flushing to the file writer.
Since it's buffered, it might take time before changes appear in the
target file. Flush every second to avoid this. The period is hard coded
for now, I don't really see much need for this to be configurable.

I considered rotating if the flush failes, but it's problematic since
that will reset the buffer. Better to keep it around and try again.