Add an elasticsearch writer to plogd as a plugin.
For the plugin to work a few structs had to be moved to a common
package. I took that opportunity to clean them up slightly and avoid
using Write() for things that are not io.Writer compatible.

The plugin is created as a submodule. This allows me to have a lot more
depenencies in its separate go.mod while still keeping them low for the
main sebase one.

I decided to start using slog in plogd so some minor changes where
needed there as well, in this case to be able to plug it in to the
elastic package.